+++ src/main.tsx
@@ -1,9 +1,9 @@
-import '@/styles/style.scss'
 import App from '@/app/App'
+import '@/styles/style.scss'
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
-import ErrorBoundary from '@/app/providers/errorBoundary/errorBoundary'
+import ErrorBoundary from '@/app/providers/ErrorBoundary/ErrorBoundary'
 
 const rootElement = document.createElement('div')
 rootElement.id = 'root'
